refactor of db creation
