+++ lib/utils.js
@@ -1,110 +1,55 @@
-const Crypto = require('crypto');
+const hashSync = require('bcryptjs').hashSync;
+const randomBytes = require('crypto').randomBytes;
 const Deepcopy = require('deepcopy');
 const Fs = require('fs');
+
+const saltRounds = 10;
 
 const {
   loweringsTable,
   cruisesTable
 } = require('../config/db_constants');
 
-const randomString = (length, chars) => {
+const addEventRecordIDs = async (request, records) => {
 
-  if (!chars) {
-    throw new Error('Argument \'chars\' is undefined');
-  }
+  const db = request.mongo.db;
 
-  const charsLength = chars.length;
-  if (charsLength > 256) {
-    throw new Error('Argument \'chars\' should not have more than 256 characters'
-      + ', otherwise unpredictability will be broken');
-  }
+  const new_results = await records.map(async (doc) => {
 
-  const randomBytes = Crypto.randomBytes(length);
-  const result = new Array(length);
+    const cruise_lowering_query = {};
 
-  let cursor = 0;
-  for (let i = 0; i < length; ++i) {
-    cursor += randomBytes[i];
-    result[i] = chars[cursor % charsLength];
-  }
+    // time bounds based on event start/stop times
+    cruise_lowering_query.$and = [{ start_ts: { $lte: doc.ts } }, { stop_ts: { $gte: doc.ts } }];
 
-  return result.join('');
-};
+    try {
+      const event_cruise = await db.collection(cruisesTable).findOne(cruise_lowering_query);
 
-const randomAsciiString = (length) => {
-
-  return randomString(length,
-    'abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789');
-};
-
-const rmDir = (dirPath) => {
-
-  try {
-    const files = Fs.readdirSync(dirPath);
-
-    if (files.length > 0) {
-      for (let i = 0; i < files.length; ++i) {
-        const filePath = dirPath + '/' + files[i];
-        if (Fs.statSync(filePath).isFile()) {
-          Fs.unlinkSync(filePath);
-        }
-        else {
-          rmDir(filePath);
-        }
+      if (event_cruise) {
+        doc.cruise_id = event_cruise.cruise_id;
       }
     }
-  }
-  catch (err) {
-    console.log(err);
-    throw err;
-  }
+    catch (err) {
+      console.error('ERROR:', err);
+    }
 
-  try {
-    Fs.rmdirSync(dirPath);
-  }
-  catch (err) {
-    console.log(err);
-    throw err;
-  }
-};
+    try {
+      const event_lowering = await db.collection(loweringsTable).findOne(cruise_lowering_query);
 
-const mvFilesToDir = (sourceDirPath, destDirPath) => {
-
-  try {
-    const files = Fs.readdirSync(sourceDirPath);
-    if (files.length > 0) {
-      for (let i = 0; i < files.length; ++i) {
-        const sourceFilePath = sourceDirPath + '/' + files[i];
-        const destFilePath = destDirPath + '/' + files[i];
-        if (Fs.statSync(sourceFilePath).isFile()) {
-          try {
-            Fs.renameSync(sourceFilePath, destFilePath );
-          }
-          catch (error) {
-            if (error.code === 'EXDEV') {
-              Fs.copyFileSync(sourceFilePath, destFilePath );
-              Fs.unlinkSync(sourceFilePath );
-            }
-          }
-        }
-        else {
-          mvFilesToDir(sourceFilePath, destFilePath);
-        }
+      if (event_lowering) {
+        doc.lowering_id = event_lowering.lowering_id;
       }
     }
-  }
-  catch (err) {
-    console.log(err);
-    throw err;
-  }
+    catch (err) {
+      console.error('ERROR:', err);
+    }
 
-  try {
-    Fs.rmdirSync(sourceDirPath);
-  }
-  catch (err) {
-    console.log(err);
-    throw err;
-  }
+    return doc;
+  });
+
+  await Promise.all(new_results);
+
+  return records;
+
 };
 
 const arrayMove = (arr, old_index, new_index) => {
@@ -118,78 +63,6 @@ const arrayMove = (arr, old_index, new_index) => {
 
   arr.splice(new_index, 0, arr.splice(old_index, 1)[0]);
   return arr; // for testing
-};
-
-const flattenEventObjs = (event_objs) => {
-
-  const flat_events = event_objs.map((event) => {
-
-    const copied_event = Deepcopy(event);
-    copied_event.id = event.id; // weird bug where the original objectID doesn't copy.
-    copied_event.event_free_text = copied_event.event_free_text.replace('\n','\\n');
-
-    let enumerator = 0;
-    if (copied_event.aux_data) {
-
-      copied_event.aux_data.map((data) => {
-
-        data.data_array.map((data2) => {
-
-          const elementName = `${data.data_source}.${data2.data_name}`;
-          const elementUOM = `${data.data_source}.${data2.data_name}`;
-
-          if (!(elementName + '_value' in copied_event)) {
-            copied_event[elementName + '_value'] = data2.data_value;
-            copied_event[elementUOM + '_uom'] = data2.data_uom;
-          }
-          else {
-            enumerator = 2;
-            while (enumerator > 1) {
-              if (!(elementName + '_' + enumerator  + '_value' in copied_event)) {
-                copied_event[elementName + '_' + enumerator + '_value'] = data2.data_value;
-                copied_event[elementUOM + '_' + enumerator + '_uom'] = data2.data_uom;
-                enumerator = 1;
-              }
-              else {
-                enumerator++;
-              }
-            }
-          }
-        });
-      });
-      delete copied_event.aux_data;
-    }
-
-    enumerator = 0;
-    copied_event.event_options.map((data) => {
-
-      const elementName = `event_option.${data.event_option_name}`;
-      if (!(elementName in copied_event)) {
-        copied_event[elementName] = data.event_option_value;
-      }
-      else {
-        enumerator = 2;
-        while (enumerator > 1) {
-          if (!(elementName + '_' + enumerator in copied_event)) {
-            copied_event[elementName + '_' + enumerator] = data.event_option_value;
-            enumerator = 1;
-          }
-          else {
-            enumerator++;
-          }
-        }
-      }
-    });
-
-    delete copied_event.event_options;
-
-    copied_event.ts = copied_event.ts.toISOString();
-    copied_event.id = copied_event.id.toString('hex');
-    copied_event.event_free_text = copied_event.event_free_text.replace(/"/g, '"');
-    return copied_event;
-  });
-
-  return flat_events;
 };
 
 const buildEventCSVHeaders = (flat_events) => {
@@ -301,56 +174,192 @@ const buildEventsQuery = (request, start_ts = new Date('1970-01-01T00:00:00.000Z
   return query;
 };
 
-const addEventRecordIDs = async (request, records) => {
+const flattenEventObjs = (event_objs) => {
 
-  const db = request.mongo.db;
+  const flat_events = event_objs.map((event) => {
 
-  const new_results = await records.map(async (doc) => {
+    const copied_event = Deepcopy(event);
+    copied_event.id = event.id; // weird bug where the original objectID doesn't copy.
+    copied_event.event_free_text = copied_event.event_free_text.replace('\n','\\n');
 
-    const cruise_lowering_query = {};
+    let enumerator = 0;
+    if (copied_event.aux_data) {
 
-    // time bounds based on event start/stop times
-    cruise_lowering_query.$and = [{ start_ts: { $lte: doc.ts } }, { stop_ts: { $gte: doc.ts } }];
+      copied_event.aux_data.map((data) => {
 
-    try {
-      const event_cruise = await db.collection(cruisesTable).findOne(cruise_lowering_query);
+        data.data_array.map((data2) => {
 
-      if (event_cruise) {
-        doc.cruise_id = event_cruise.cruise_id;
+          const elementName = `${data.data_source}.${data2.data_name}`;
+          const elementUOM = `${data.data_source}.${data2.data_name}`;
+
+          if (!(elementName + '_value' in copied_event)) {
+            copied_event[elementName + '_value'] = data2.data_value;
+            copied_event[elementUOM + '_uom'] = data2.data_uom;
+          }
+          else {
+            enumerator = 2;
+            while (enumerator > 1) {
+              if (!(elementName + '_' + enumerator  + '_value' in copied_event)) {
+                copied_event[elementName + '_' + enumerator + '_value'] = data2.data_value;
+                copied_event[elementUOM + '_' + enumerator + '_uom'] = data2.data_uom;
+                enumerator = 1;
+              }
+              else {
+                enumerator++;
+              }
+            }
+          }
+        });
+      });
+      delete copied_event.aux_data;
+    }
+
+    enumerator = 0;
+    copied_event.event_options.map((data) => {
+
+      const elementName = `event_option.${data.event_option_name}`;
+      if (!(elementName in copied_event)) {
+        copied_event[elementName] = data.event_option_value;
       }
-    }
-    catch (err) {
-      console.error('ERROR:', err);
-    }
-
-    try {
-      const event_lowering = await db.collection(loweringsTable).findOne(cruise_lowering_query);
-
-      if (event_lowering) {
-        doc.lowering_id = event_lowering.lowering_id;
+      else {
+        enumerator = 2;
+        while (enumerator > 1) {
+          if (!(elementName + '_' + enumerator in copied_event)) {
+            copied_event[elementName + '_' + enumerator] = data.event_option_value;
+            enumerator = 1;
+          }
+          else {
+            enumerator++;
+          }
+        }
       }
-    }
-    catch (err) {
-      console.error('ERROR:', err);
-    }
+    });
 
-    return doc;
+    delete copied_event.event_options;
+
+    copied_event.ts = copied_event.ts.toISOString();
+    copied_event.id = copied_event.id.toString('hex');
+    copied_event.event_free_text = copied_event.event_free_text.replace(/"/g, '"');
+    return copied_event;
   });
 
-  await Promise.all(new_results);
+  return flat_events;
+};
 
-  return records;
+const hashedPassword = (password) => {
+
+  return hashSync( password, saltRounds );
 
 };
 
+const mvFilesToDir = (sourceDirPath, destDirPath) => {
+
+  try {
+    const files = Fs.readdirSync(sourceDirPath);
+    if (files.length > 0) {
+      for (let i = 0; i < files.length; ++i) {
+        const sourceFilePath = sourceDirPath + '/' + files[i];
+        const destFilePath = destDirPath + '/' + files[i];
+        if (Fs.statSync(sourceFilePath).isFile()) {
+          try {
+            Fs.renameSync(sourceFilePath, destFilePath );
+          }
+          catch (error) {
+            if (error.code === 'EXDEV') {
+              Fs.copyFileSync(sourceFilePath, destFilePath );
+              Fs.unlinkSync(sourceFilePath );
+            }
+          }
+        }
+        else {
+          mvFilesToDir(sourceFilePath, destFilePath);
+        }
+      }
+    }
+  }
+  catch (err) {
+    console.log(err);
+    throw err;
+  }
+
+  try {
+    Fs.rmdirSync(sourceDirPath);
+  }
+  catch (err) {
+    console.log(err);
+    throw err;
+  }
+};
+
+const randomAsciiString = (length) => {
+
+  return randomString(length,
+    'abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789');
+};
+
+const randomString = (length, chars) => {
+
+  if (!chars) {
+    throw new Error('Argument \'chars\' is undefined');
+  }
+
+  const charsLength = chars.length;
+  if (charsLength > 256) {
+    throw new Error('Argument \'chars\' should not have more than 256 characters'
+      + ', otherwise unpredictability will be broken');
+  }
+
+  const random_bytes = randomBytes(length);
+  const result = new Array(length);
+
+  let cursor = 0;
+  for (let i = 0; i < length; ++i) {
+    cursor += random_bytes[i];
+    result[i] = chars[cursor % charsLength];
+  }
+
+  return result.join('');
+};
+
+const rmDir = (dirPath) => {
+
+  try {
+    const files = Fs.readdirSync(dirPath);
+
+    if (files.length > 0) {
+      for (let i = 0; i < files.length; ++i) {
+        const filePath = dirPath + '/' + files[i];
+        if (Fs.statSync(filePath).isFile()) {
+          Fs.unlinkSync(filePath);
+        }
+        else {
+          rmDir(filePath);
+        }
+      }
+    }
+  }
+  catch (err) {
+    console.log(err);
+    throw err;
+  }
+
+  try {
+    Fs.rmdirSync(dirPath);
+  }
+  catch (err) {
+    console.log(err);
+    throw err;
+  }
+};
 
 module.exports = {
-  randomString,
-  randomAsciiString,
-  rmDir,
-  mvFilesToDir,
-  flattenEventObjs,
+  addEventRecordIDs,
   buildEventCSVHeaders,
   buildEventsQuery,
-  addEventRecordIDs
+  flattenEventObjs,
+  hashedPassword,
+  mvFilesToDir,
+  randomAsciiString,
+  randomString,
+  rmDir
 };
