+++ plugins/db_custom_vars.js
@@ -2,57 +2,63 @@ const {
   customVarsTable
 } = require('../config/db_constants');
 
+const {
+  custom_vars_init_data
+} = require('../lib/db_init_data');
+
+let env = process.env.NODE_ENV || 'development';
+env = (env === 'test') ? 'development' : env;
+env = (env === 'debug') ? 'production' : env;
+
 exports.plugin = {
   name: 'db_populate_custom_vars',
   dependencies: ['hapi-mongodb'],
   register: async (server, options) => {
 
     const db = server.mongo.db;
-    const ObjectID = server.mongo.ObjectID;
 
-    const init_data = [
-      {
-        _id: ObjectID('59810167212b348aed7fa9f5'),
-        custom_var_name: 'asnapStatus',
-        custom_var_value: 'Off'
-      },
-      {
-        _id: ObjectID('59810167212b348aed7fa9f6'),
-        custom_var_name: 'freeSpaceInBytes',
-        custom_var_value: '0'
-      },
-      {
-        _id: ObjectID('59810167212b348aed7fa9f7'),
-        custom_var_name: 'freeSpacePercentage',
-        custom_var_value: '0'
-      }
-
-    ];
-
-    console.log('Searching for Custom Variable Collection');
     try {
+      console.log('Searching for Custom Vars Collection');
       const result = await db.listCollections({ name: customVarsTable }).toArray();
-      if (result.length > 0) {
-        console.log('Collection already exists... we\'re done here.');
-        return;
+      if (result.length > 0 ) {
+        if (env === 'production' ) {
+          console.log('Custom Vars Collection already exists... we\'re done here.');
+          return;
+        }
+
+        try {
+          console.log('Custom Vars Collection already exists... we\'re dropping it.');
+          await db.dropCollection(customVarsTable);
+        }
+        catch (err) {
+          console.log('DROP ERROR:', err.code);
+          throw (err);
+        }
       }
     }
     catch (err) {
-      console.log('ERROR:', err.code);
+      console.log('LIST ERROR:', err.code);
       throw (err);
     }
 
+    console.log('Creating Custom Vars Collection');
     try {
-      console.log('Creating Custom Variable Collection');
       const collection = await db.createCollection(customVarsTable);
 
-      console.log('Populating Custom Variable Collection');
-      await collection.insertMany(init_data);
+      try {
+        console.log('Populating Custom Vars Collection');
+        await collection.insertMany(custom_vars_init_data);
 
+      }
+      catch (err) {
+        console.log('INSERT ERROR:', err.code);
+        throw (err);
+      }
     }
     catch (err) {
-      console.log('ERROR:', err.code);
+      console.log('CREATE ERROR:', err.code);
       throw (err);
     }
+
   }
 };
